scraper-service/requirements.txt
crawl4ai>=0.1.0
playwright>=1.44.0
aiohttp
pandas
openai
requests
pyyaml
